requirements.txt
# Core dependencies
numpy>=1.21.6
torch>=1.12.0
torchaudio>=0.12.0
transformers>=4.20.0
pydub>=0.25.1
librosa>=0.9.2
soundfile>=0.10.3
scipy>=1.7.3
matplotlib>=3.5.0

# Script parsing
markdown>=3.3.0
beautifulsoup4>=4.10.0
regex>=2022.3.0

# Audio processing
audiomentations>=0.20.0
pedalboard>=0.5.0
wave>=0.0.2

# TTS/Voice models
huggingface_hub>=0.10.0
onnxruntime>=1.11.0
# onnxruntime-gpu>=1.11.0
onnxruntime>=1.11.0
speechbrain>=0.5.12

# Testing
pytest>=7.0.0
pytest-cov>=2.12.0

# Utilities
tqdm>=4.64.0
click>=8.1.0
PyYAML>=6.0.0
joblib>=1.1.0